some fixes + added clear option

+++ git_utils.cmake
@@ -2,7 +2,7 @@ cmake_minimum_required(VERSION 3.5)
 
 function(get_latest_tag)
         # Parse args
-        set(ONE_VALUE_ARGS PROFILE_NAME REPOSITORY_NAME OUTPUT_VARIABLE)
+        set(ONE_VALUE_ARGS PROFILE_NAME REPOSITORY_NAME CLEAR OUTPUT_VARIABLE)
         set(MULTI_VALUE_ARGS)
         cmake_parse_arguments(ARGS "" "${ONE_VALUE_ARGS}" "${MULTI_VALUE_ARGS}" ${ARGN})
 
@@ -12,17 +12,23 @@ function(get_latest_tag)
         # Temporary directory name
         set(TMP_DIR tmp_${ARGS_REPOSITORY_NAME})
 
-        # Clone repo with lowest depth possible
-        execute_process(
-                COMMAND git clone --depth 1 --no-checkout ${GITHUB_URL} ${TMP_DIR}
-                WORKING_DIRECTORY ${CMAKE_BINARY_DIR}
-        )
+        # False if CLEAR was false in a previous call.
+        if (NOT EXISTS ${TMP_DIR})
+                # Clone repo with lowest depth possible
+                execute_process(
+                        COMMAND git clone --depth 1 --no-checkout ${GITHUB_URL} ${TMP_DIR}
+                        WORKING_DIRECTORY ${CMAKE_BINARY_DIR}
+                )
 
-        # Fetch all tags
-        execute_process(
-                COMMAND git fetch --tags --depth 1
-                WORKING_DIRECTORY "${CMAKE_BINARY_DIR}/${TMP_DIR}"
-        )
+                # Fetch all tags
+                execute_process(
+                        COMMAND git fetch --tags --depth 1
+                        WORKING_DIRECTORY "${CMAKE_BINARY_DIR}/${TMP_DIR}"
+                )
+        else ()
+                message(STATUS "git_utils::get_latest_tag() temporary directory for repository '${ARGS_REPOSITORY_NAME}' "
+                        "already exists, not fetching.")
+        endif ()
 
         # Sort tags by creation date
         execute_process(
@@ -33,13 +39,14 @@ function(get_latest_tag)
         )
 
         # Delete downloaded content
-        file(REMOVE_RECURSE "${CMAKE_BINARY_DIR}/${TMP_DIR}")
+        if (CLEAR)
+                file(REMOVE_RECURSE "${CMAKE_BINARY_DIR}/${TMP_DIR}")
+        endif ()
 
         # Checking if tag list was obtained correctly
         if (NOT TAG_LIST)
                 message(FATAL_ERROR "Failed to obtain tag list.")
         endif ()
-        message(STATUS ${TAG_LIST})
 
         # Get the latest tag from the tag list
         string(REGEX MATCH "([^ \n]+)" TAG_NAME ${TAG_LIST})
